Move, rename, refactor tests
