+++ tests/functional/spec/features/placeholder/placeholder_date_difference.spec.js
@@ -1,11 +1,13 @@
-import AgeBlockYearPage from "../generated_pages/difference_in_years/age-block.page";
-import AgeBlockMonthYearPage from "../generated_pages/difference_in_years_month_year/age-block.page";
-import AgeBlockDayMonthYearRangePage from "../generated_pages/difference_in_years_range/date-block.page";
-import AgeBlockMonthYearRangePage from "../generated_pages/difference_in_years_month_year_range/date-block.page";
+import AgeBlockYearPage from "../../../generated_pages/placeholder_difference_in_years/age-block.page";
+import AgeBlockMonthYearPage from "../../../generated_pages/placeholder_difference_in_years_month_year/age-block.page";
+import AgeBlockDayMonthYearRangePage from "../../../generated_pages/placeholder_difference_in_years_range/date-block.page";
+import AgeTestDayMonthYearRangePage from "../../../generated_pages/placeholder_difference_in_years_range/age-test.page";
+import AgeBlockMonthYearRangePage from "../../../generated_pages/placeholder_difference_in_years_month_year_range/date-block.page";
+import AgeTestMonthYearRangePage from "../../../generated_pages/placeholder_difference_in_years_month_year_range/age-test.page";
 
 describe("Difference check (years)", () => {
   before("Load the survey", () => {
-    browser.openQuestionnaire("test_difference_in_years.json");
+    browser.openQuestionnaire("test_placeholder_difference_in_years.json");
   });
 
   it("Given a day, month and year answer is provided for a date question then the age in years should be calculated and displayed on the page ", () => {
@@ -21,7 +23,7 @@ describe("Difference check (years)", () => {
 
 describe("Difference check (months and years)", () => {
   before("Load the survey", () => {
-    browser.openQuestionnaire("test_difference_in_years_month_year.json");
+    browser.openQuestionnaire("test_placeholder_difference_in_years_month_year.json");
   });
 
   it("Given a month and year answer is provided for a date question then the difference in years should be calculated and displayed on the page ", () => {
@@ -37,7 +39,7 @@ describe("Difference check (months and years)", () => {
 
 describe("Difference check (months and years range)", () => {
   before("Load the survey", () => {
-    browser.openQuestionnaire("test_difference_in_years_month_year_range.json");
+    browser.openQuestionnaire("test_placeholder_difference_in_years_month_year_range.json");
   });
 
   it("Given a month and year answers 'from' and 'to' are provided for a date question then the difference in years should be calculated and displayed on the page ", () => {
@@ -48,13 +50,13 @@ describe("Difference check (months and years range)", () => {
 
     $(AgeBlockMonthYearRangePage.submit()).click();
 
-    expect($("body").getText()).to.have.string("You were out of the UK for 1 year. Is this correct?");
+    expect($(AgeTestMonthYearRangePage.heading()).getText()).to.have.string("You were out of the UK for 1 year. Is this correct?");
   });
 });
 
 describe("Difference check (years range)", () => {
   before("Load the survey", () => {
-    browser.openQuestionnaire("test_difference_in_years_range.json");
+    browser.openQuestionnaire("test_placeholder_difference_in_years_range.json");
   });
 
   it("Given a day, month and year answers 'from' and 'to' are provided for a date question then the difference in years should be calculated and displayed on the page ", () => {
@@ -68,6 +70,6 @@ describe("Difference check (years range)", () => {
 
     $(AgeBlockDayMonthYearRangePage.submit()).click();
 
-    expect($("body").getText()).to.have.string("You were out of the UK for 1 year. Is this correct?");
+    expect($(AgeTestDayMonthYearRangePage.heading()).getText()).to.have.string("You were out of the UK for 1 year. Is this correct?");
   });
 });
